requirements.txt
click
psycopg2-binary
pytest
pytest-mock
pytest-cov
black
pre-commit